Adding gatsby-plugin keyword in package.json
Hello 👋

As part of gatsbyjs/gatsby#14013, I would like to add gatsby and gatsby-plugin keywords to the package.json file of your source plugin.

It is documented in https://www.gatsbyjs.org/contributing/submit-to-plugin-library/ if you would like to know more about it.

If you accept to merge this PR, could you also publish a patch version of your source plugin on NPM? Let me know if I can assist in any way with this. 🙂

Thank you very much!